requirements.txt
streamlit
pandas
openpyxl
google-generativeai
matplotlib
seaborn
